Toggling a book does not refresh the whole list
Previously toggling a book on the Enabled Books page would refresh the entire flatlist, causing the user to be presented with the top of the list of books again, even if they wanted to be at the bottom of the list.

Toggling a book no longer refreshes the whole list.