Changed longpress and shortpress on Enabled Books
Based on user feedback, on the Enabled Books page, the short press now handles book toggles, and the long press handles book expansion.

This may be updated later so that the user can decide which press does what based on an account setting, but this will likely not be addressed until the backend of the app has been refactored.