updated gitignore to include python scripts
I've added to the repo a python script that'll help me update the json Bible data so that the Verses screen will account for multiple correct answers. As this script will theoretically only need to be used once and won't be part of the app, I figured it doesn't need to be included in the GitHub repo. Contact me if you want the script.